Split google map components
