+++ app/frontend/containers/MapDetailContainer.js
@@ -11,26 +11,13 @@ import getMapBasePosition from '../actions/getMapBasePosition';
 import requestMapBase from '../actions/requestMapBase';
 import fetchSpots from '../actions/fetchSpots';
 import fetchCollaborators from '../actions/fetchCollaborators';
-import openSpotCard from '../actions/openSpotCard';
 import clearMapState from '../actions/clearMapState';
-import gMapMounted from '../actions/gMapMounted';
-import mapZoomChanged from '../actions/mapZoomChanged';
-import mapCenterChanged from '../actions/mapCenterChanged';
 import fetchMapReviews from '../actions/fetchMapReviews';
-import selectSpot from '../actions/selectSpot';
 
 const mapStateToProps = state => {
   return {
-    gMap: state.gMap.gMap,
-    currentMap: state.mapDetail.currentMap,
-    defaultCenter: state.gMap.defaultCenter,
-    defaultZoom: state.gMap.defaultZoom,
-    center: state.gMap.center,
-    zoom: state.gMap.zoom,
-    spots: state.gMap.spots,
-    currentPosition: state.gMap.currentPosition,
-    directions: state.gMap.directions,
     large: state.shared.large,
+    currentMap: state.mapDetail.currentMap,
     currentSpot: state.spotCard.currentSpot,
     spotCardOpen: state.spotCard.spotCardOpen,
     mapReviews: state.mapSummary.mapReviews,
@@ -42,6 +29,15 @@ const mapDispatchToProps = (dispatch, ownProps) => {
   return {
     updatePageTitle: (mapName = undefined) => {
       dispatch(updatePageTitle(mapName ? mapName : 'Map'));
+    },
+
+    initCenter: async map => {
+      if (map.base.place_id) {
+        dispatch(getMapBasePosition(map.base.lat, map.base.lng));
+        dispatch(requestMapBase());
+      } else {
+        dispatch(requestCurrentPosition());
+      }
     },
 
     fetchMap: async () => {
@@ -56,15 +52,6 @@ const mapDispatchToProps = (dispatch, ownProps) => {
         dispatch(push('/maps'));
       } else {
         dispatch(openToast('Failed to fetch Map.'));
-      }
-    },
-
-    initCenter: async map => {
-      if (map.base.place_id) {
-        dispatch(getMapBasePosition(map.base.lat, map.base.lng));
-        dispatch(requestMapBase());
-      } else {
-        dispatch(requestCurrentPosition());
       }
     },
 
@@ -99,23 +86,6 @@ const mapDispatchToProps = (dispatch, ownProps) => {
 
     handleCreateReviewClick: () => {
       dispatch(openPlaceSelectDialog());
-    },
-
-    onSpotMarkerClick: async spot => {
-      dispatch(selectSpot(spot));
-      dispatch(openSpotCard());
-    },
-
-    onMapMounted: map => {
-      dispatch(gMapMounted(map));
-    },
-
-    onZoomChanged: zoom => {
-      dispatch(mapZoomChanged(zoom));
-    },
-
-    onCenterChanged: center => {
-      dispatch(mapCenterChanged({ lat: center.lat(), lng: center.lng() }));
     },
 
     handleUnmount: () => {
